store/tikv: move mocktikv into tikv (#24099)
